Refactored changing filenames and directory structures
