+++ analysis/min_rsu_for_specific_num_ligs/src/num2_thetastep10.py
@@ -1,6 +1,6 @@
 import os
 
-from rsuanalyzer.ranker.min_rsu import calc_min_rsu_for_thetas
+from rsuanalyzer.analyze_rsu.min_rsu import calc_min_rsu_for_thetas
 
 NUM_LIGS = 2
 THETA_STEP = 10
