+++ analysis/rank_for_specific_num_ligs_and_theta/src/num3_theta34.py
@@ -1,6 +1,6 @@
 import os
 
-from rsuanalyzer.ranker.make_rank_table import make_rank_table
+from rsuanalyzer.analyze_rsu.make_rank_table import make_rank_table
 
 NUM_LIGS = 3
 THETA = 34
