+++ analysis/rsu_of_specific_rings/src/rings_in_Pd6L4_Pd9L6_Pd12L8.py
@@ -1,7 +1,7 @@
 import os
 from dataclasses import dataclass
 
-from rsuanalyzer.core.rsu import calc_rsu
+from rsuanalyzer.calc_rsu.rsu import calc_rsu
 
 
 @dataclass
