+++ analysis/rsu_vs_theta_for_specific_ring/src/syns2_delta90.py
@@ -7,7 +7,7 @@ import os
 
 import pandas as pd
 
-from rsuanalyzer.core.rsu import calc_rsu
+from rsuanalyzer.calc_rsu.rsu import calc_rsu
 
 CONF_ID_OF_RING = "RRFBRLBBRRFBRLBB"
 THETAS = range(0, 91, 1)
