+++ chainvisualizer/chain.py
@@ -1,8 +1,9 @@
 
 import numpy as np
 
-from rsuanalyzer.core.chain import calc_con_rot, calc_lig_ends_of_chain
-from rsuanalyzer.core.conf_id import conf_id_to_lig_and_con_types
+from rsuanalyzer.calc_rsu.chain import calc_con_rot, calc_lig_ends_in_chain
+from rsuanalyzer.calc_rsu.conf_id_to_lig_and_con_types import \
+    conf_id_to_lig_and_con_types
 
 from .lig import calc_c_positions_of_frags_in_lig
 
@@ -11,7 +12,7 @@ def calc_metal_positions(
         conf_id: str, theta: float, delta_: float
         ) -> list[np.ndarray]:
     metal_positions = [np.array([0, 0, 0])]
-    lig_ends = calc_lig_ends_of_chain(conf_id, theta, delta_)
+    lig_ends = calc_lig_ends_in_chain(conf_id, theta, delta_)
     for lig_end in lig_ends:
         x, rot = lig_end
         metal_positions.append(x)
@@ -21,7 +22,7 @@ def calc_metal_positions(
 def calc_carbon_positions(conf_id: str, theta: float, delta_: float):
     lig_types, con_types = conf_id_to_lig_and_con_types(conf_id)
 
-    lig_ends = calc_lig_ends_of_chain(conf_id, theta, delta_)
+    lig_ends = calc_lig_ends_in_chain(conf_id, theta, delta_)
     carbon_positions_of_all_ligs = [calc_c_positions_of_frags_in_lig(lig_types[0], theta)]
 
     for lig_type, con_type, prev_lig_end in zip(lig_types[1:], con_types, lig_ends[:-1]):
