+++ rsuanalyzer/calc_rsu/rsu.py
@@ -1,7 +1,4 @@
-import numpy as np
-
 from .chain import calc_chain_end_dist
-from .conf_id import ring_to_chains
 
 
 def calc_rsu(
@@ -31,7 +28,7 @@ def calc_rsu(
             "multiple of 4.")
     
     chain_length = len(conf_id_of_ring) // 4
-    conf_ids = ring_to_chains(conf_id_of_ring)
+    conf_ids = _ring_id_to_chain_ids(conf_id_of_ring)
 
     ave_chain_end_dist = sum(
         calc_chain_end_dist(conf_id, theta, delta_) 
@@ -40,3 +37,36 @@ def calc_rsu(
     return ave_chain_end_dist / chain_length
 
 
+def _ring_id_to_chain_ids(conf_id_of_ring: str) -> list[str]:
+    """
+    Convert the conformation ID of a ring to the conformation IDs of
+    the chains.
+
+    Args:
+    - conf_id_of_ring (str): The conformation ID of the ring. 
+      This should contain the same number of connection types as the 
+      number of ligands in the ring. For example, for a 2-membered ring, 
+      it would be "RRFFRLFF".
+
+    Returns:
+    - list[str]: The conformation IDs of the chains.
+
+    Examples:
+    >>> ring_to_chains("RRFFRLFF")
+    ['RRFFRL', 'RLFFRR']
+    >>> ring_to_chains("RRFFRLFFRRBF")
+    ['RRFFRLFFRR', 'RRBFRRFFRL', 'RLFFRRBFRR']
+    """
+    # Validate the input.
+    if len(conf_id_of_ring) % 4 != 0:
+        raise ValueError(
+            "The length of the conformation ID of the ring should be a "
+            "multiple of 4.")
+    
+    chain_length = len(conf_id_of_ring) // 4
+    chains = []
+    for _ in range(chain_length):
+        conf_id = conf_id_of_ring[:-2]
+        conf_id_of_ring = conf_id_of_ring[-4:] + conf_id_of_ring[:-4]
+        chains.append(conf_id)
+    return chains
